# openapi/openapi.yaml
openapi: 3.0.3
info:
  description: Self-service eSIM API. Public REST under `/api/v1/`. Staff Operations Dashboard under `/api/v1/admin/`
    (read-only). Staff internal tools under `/api/internal/`. Authenticate with JWT Bearer tokens from
    `/api/v1/auth/token/`.
  title: RoamKit API
  version: 1.0.0
servers:
- description: Staging
  url: https://api.staging.roamkit.net
- description: Production
  url: https://api.roamkit.net
...
paths:
...
  /api/v1/orgs/partner/invite-link/:
    get:
      description: Record that a legacy partner route was called. Does not change the response.
      operationId: orgs_partner_invite_link
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerInviteLink'
          description: Invite link
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: authentication_required
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: partner_access_denied or partner_invite_forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: partner_channel_disabled
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: partner_context_ambiguous
      security:
      - bearerAuth: []
      summary: Partner invite link
      tags:
      - Organizations
  /api/v1/orgs/partner/invite-link/activate/:
    post:
      description: Record that a legacy partner route was called. Does not change the response.
      operationId: orgs_partner_invite_link_activate
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerInviteLink'
          description: Invite link
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: authentication_required
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: partner_access_denied or partner_invite_forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: partner_channel_disabled
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: partner_context_ambiguous
      security:
      - bearerAuth: []
      summary: Activate partner invite link
      tags:
      - Organizations
  /api/v1/orgs/partner/invite-link/deactivate/:
    post:
      description: Record that a legacy partner route was called. Does not change the response.
      operationId: orgs_partner_invite_link_deactivate
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerInviteLink'
          description: Invite link
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: authentication_required
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: partner_access_denied or partner_invite_forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: partner_channel_disabled
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: partner_context_ambiguous
      security:
      - bearerAuth: []
      summary: Deactivate partner invite link
      tags:
      - Organizations
  /api/v1/orgs/partner/invite-link/regenerate/:
    post:
      description: Record that a legacy partner route was called. Does not change the response.
      operationId: orgs_partner_invite_link_regenerate
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerInviteLink'
          description: Invite link
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: authentication_required
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: partner_access_denied or partner_invite_forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: partner_channel_disabled
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: partner_context_ambiguous
      security:
      - bearerAuth: []
      summary: Regenerate partner invite link
      tags:
      - Organizations
  /api/v1/orgs/partner/summary/:
    get:
      description: Stored margin total, display balance, and accrual counts for the caller's single partner
        channel.
      operationId: orgs_partner_summary
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerSummary'
          description: Channel summary
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: authentication_required
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: partner_access_denied
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: partner_channel_disabled
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerChannelCode'
          description: partner_context_ambiguous
      security:
      - bearerAuth: []
      summary: Partner channel summary
      tags:
      - Organizations
...
components:
  schemas:
...
    PartnerChannelCode:
      description: Partner read error body. Exactly one code, no field names.
      properties:
        code:
          type: string
      required:
      - code
      type: object
...
    PartnerInviteLink:
      properties:
        created_at:
          format: date-time
          type: string
        is_active:
          type: boolean
        regenerated_at:
          format: date-time
          nullable: true
          type: string
        url:
          type: string
      required:
      - created_at
      - is_active
      - regenerated_at
      - url
      type: object
    PartnerSummary:
      description: Six-decimal strings for money. Counts stay integers.
      properties:
        accrual_counts:
          $ref: '#/components/schemas/PartnerSummaryCounts'
        available_balance:
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,6})?$
          type: string
        total_earned:
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,6})?$
          type: string
      required:
      - accrual_counts
      - available_balance
      - total_earned
      type: object
    PartnerSummaryCounts:
      properties:
        order:
          minimum: 0
          type: integer
        subscription:
          minimum: 0
          type: integer
        topup:
          minimum: 0
          type: integer
        total:
          minimum: 0
          type: integer
      required:
      - order
      - subscription
      - topup
      - total
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http